% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_signature_mw.R
\name{lsp_signature_mw}
\alias{lsp_signature_mw}
\title{Spatial signatures in moving window}
\usage{
lsp_signature_mw(
  x,
  window,
  neighbourhood = 4,
  ordered = FALSE,
  normalization = "pdf",
  classes = NULL,
  ...
)
}
\arguments{
\item{x}{SpatRaster}

\item{window}{Moving window matrix.}

\item{neighbourhood}{The number of directions in which cell adjacencies are considered as neighbours: 4 (rook's case) or 8 (queen's case). The default is 4.}

\item{ordered}{The type of pairs considered. Either "ordered" (TRUE) or "unordered" (FALSE). The default is FALSE.}

\item{normalization}{Should the output vector be normalized? Either "none" or "pdf". The "pdf" option normalizes a vector to sum to one. The default is "pdf".}

\item{classes}{Which classes (categories) should be analyzed? The default is NULL, which means that the classes are calculated directly from the input data and all of them are used in the calculations.}

\item{...}{Additional argument to \code{\link[terra:focalCpp]{terra::focalCpp()}}}
}
\value{
A SpatRaster with the spatial signatures of each cell in many layers.
}
\description{
Calculates the co-occurrence vector (cove) spatial signature in a moving window based on a provided categorical raster and a window matrix.
}
\examples{
library(terra)
r = rast(matrix(c(rep(1, 23), rep(NA, 3), rep(2, 23)), ncol = 7))
plot(r)
w = matrix(1, 3, 3)
result = lsp_signature_mw(x = r, window = w)
result
}
