% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifmw-package.R
\docType{package}
\name{motifmw-package}
\alias{motifmw}
\alias{motifmw-package}
\title{motifmw: Local Pattern Analysis in Moving Window}
\description{
Describes spatial patterns of categorical raster data for any defined moving window. Patterns are described quantitatively using built-in signatures based on a co-occurrence matrix.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://jakubnowosad.com/motifmw/}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
